changed the defaults of the trigger to match current realdaq
git-svn-id: http://code.icecube.wisc.edu/svn/projects/sim-services/releases/V00-00-01@21139 16731396-06f5-0310-8873-f7f720988828

+++ public/sim-services/sim-source/I3MCSourceServiceFactory.h
@@ -36,7 +36,7 @@ class I3MCSourceServiceFactory
 
   int configID_;
   int threshold_;
-  int timeWindow_;
+  double timeWindow_;
 
   std::string calServiceName_;
   std::string statusServiceName_;
